+++ packages/remix-dev/cli/run.ts
@@ -48,7 +48,7 @@ ${colors.logoBlue("R")} ${colors.logoGreen("E")} ${colors.logoYellow(
     --http-host         HTTP(S) host for the dev server. Default: localhost
     --http-port         HTTP(S) port for the dev server. Default: any open port
     --no-restart        Do not restart the app server when rebuilds occur.
-    --websocket-port    Websocket port for the dev server. Default: any open port
+    --websocket-port    WebSocket port for the dev server. Default: any open port
   \`init\` Options:
     --no-delete         Skip deleting the \`remix.init\` script
   \`routes\` Options:
@@ -226,7 +226,7 @@ export async function run(argv: string[] = process.argv.slice(2)) {
     delete flags["http-port"];
   }
   if (flags["websocket-port"]) {
-    flags.websocketPort = flags["websocket-port"];
+    flags.webSocketPort = flags["websocket-port"];
     delete flags["websocket-port"];
   }