+++ packages/remix-dev/compiler/cancel.ts
@@ -1,0 +1,7 @@
+export const CANCEL_PREFIX = "remix-compile-cancel";
+
+export class Cancel extends Error {
+  constructor(message: string) {
+    super(`${CANCEL_PREFIX}: ${message}`);
+  }
+}