+++ packages/remix-dev/compiler/js/plugins/routes_unstable.ts
@@ -12,7 +12,7 @@ import { processMDX } from "../../plugins/mdx";
 
 const serverOnlyExports = new Set(["action", "loader"]);
 
-let removeServerExports = (onLoader: (loader: string) => void) =>
+let removeServerExports = () =>
   Transform.create(({ types: t }) => {
     return {
       visitor: {
@@ -38,20 +38,12 @@ let removeServerExports = (onLoader: (loader: string) => void) =>
           if (t.isFunctionDeclaration(node.declaration)) {
             let name = node.declaration.id?.name;
             if (!name) return;
-            if (name === "loader") {
-              let { code } = generate(node);
-              onLoader(code);
-            }
             if (serverOnlyExports.has(name)) return path.remove();
           }
           if (t.isVariableDeclaration(node.declaration)) {
             let declarations = node.declaration.declarations.filter((d) => {
               let name = t.isIdentifier(d.id) ? d.id.name : undefined;
               if (!name) return false;
-              if (name === "loader") {
-                let { code } = generate(node);
-                onLoader(code);
-              }
               return !serverOnlyExports.has(name);
             });
             if (declarations.length === 0) return path.remove();
@@ -72,8 +64,7 @@ let removeServerExports = (onLoader: (loader: string) => void) =>
  */
 export function browserRouteModulesPlugin(
   { config, options }: Context,
-  routeModulePaths: Map<string, string>,
-  onLoader: (filename: string, code: string) => void
+  routeModulePaths: Map<string, string>
 ): esbuild.Plugin {
   return {
     name: "browser-route-modules",
@@ -99,6 +90,8 @@ export function browserRouteModulesPlugin(
         };
       });
 
+      let cache = new Map();
+
       build.onLoad(
         { filter: /.*/, namespace: "browser-route-module" },
         async (args) => {
@@ -117,9 +110,7 @@ export function browserRouteModulesPlugin(
               return mdxResult;
             }
 
-            let transform = removeServerExports((loader: string) =>
-              onLoader(routeFile, loader)
-            );
+            let transform = removeServerExports();
             mdxResult.contents = transform(
               mdxResult.contents,
               // Trick babel into allowing JSX syntax.
@@ -130,28 +121,34 @@ export function browserRouteModulesPlugin(
 
           let sourceCode = fs.readFileSync(routeFile, "utf8");
 
-          let transform = removeServerExports((loader: string) =>
-            onLoader(routeFile, loader)
-          );
-          let contents = transform(sourceCode, routeFile);
+          let value = cache.get(file);
+          if (!value || value.sourceCode !== sourceCode) {
+            let transform = removeServerExports();
+            let contents = transform(sourceCode, routeFile);
 
-          if (options.mode === "development" && config.future.unstable_dev) {
-            contents = await applyHMR(
-              contents,
-              {
-                ...args,
-                path: routeFile,
+            if (options.mode === "development" && config.future.unstable_dev) {
+              contents = await applyHMR(
+                contents,
+                {
+                  ...args,
+                  path: routeFile,
+                },
+                config,
+                !!build.initialOptions.sourcemap
+              );
+            }
+            value = {
+              sourceCode,
+              output: {
+                contents,
+                loader: getLoaderForFile(routeFile),
+                resolveDir: path.dirname(routeFile),
               },
-              config,
-              !!build.initialOptions.sourcemap
-            );
+            };
+            cache.set(file, value);
           }
 
-          return {
-            contents,
-            loader: getLoaderForFile(routeFile),
-            resolveDir: path.dirname(routeFile),
-          };
+          return value.output;
         }
       );
     },