+++ packages/remix-dev/compiler/server/plugins/entry.ts
@@ -51,9 +51,9 @@ ${Object.keys(config.routes)
   export const publicPath = ${JSON.stringify(config.publicPath)};
   export const entry = { module: entryServer };
   ${
-    options.devWebsocketPort
+    options.devWebSocketPort
       ? `export const dev = ${JSON.stringify({
-          websocketPort: options.devWebsocketPort,
+          websocketPort: options.devWebSocketPort,
         })}`
       : ""
   }