+++ packages/remix-dev/devServer_unstable/index.ts
@@ -8,12 +8,13 @@ import express from "express";
 import * as Channel from "../channel";
 import { type Manifest } from "../manifest";
 import * as Compiler from "../compiler";
-import { readConfig, type RemixConfig } from "../config";
+import { type RemixConfig } from "../config";
 import { loadEnv } from "./env";
 import * as Socket from "./socket";
 import * as HMR from "./hmr";
 import { warnOnce } from "../warnOnce";
 import { detectPackageManager } from "../cli/detectPackageManager";
+import * as HDR from "./hdr";
 
 type Origin = {
   scheme: string;
@@ -23,25 +24,12 @@ type Origin = {
 
 let stringifyOrigin = (o: Origin) => `${o.scheme}://${o.host}:${o.port}`;
 
-let patchPublicPath = (
-  config: RemixConfig,
-  devHttpOrigin: Origin
-): RemixConfig => {
-  // set public path to point to dev server
-  // so that browser asks the dev server for assets
-  return {
-    ...config,
-    // dev server has its own origin, to `/build/` path will not cause conflicts with app server routes
-    publicPath: stringifyOrigin(devHttpOrigin) + "/build/",
-  };
-};
-
 let detectBin = async (): Promise<string> => {
   let pkgManager = detectPackageManager() ?? "npm";
   if (pkgManager === "npm") {
     // npm v9 removed the `bin` command, so have to use `prefix`
     let { stdout } = await execa(pkgManager, ["prefix"]);
-    return stdout.trim() + "/node_modules/.bin";
+    return path.join(stdout.trim(), "node_modules", ".bin");
   }
   let { stdout } = await execa(pkgManager, ["bin"]);
   return stdout.trim();
@@ -54,12 +42,12 @@ export let serve = async (
     httpScheme: string;
     httpHost: string;
     httpPort: number;
-    websocketPort: number;
+    webSocketPort: number;
     restart: boolean;
   }
 ) => {
   await loadEnv(initialConfig.rootDirectory);
-  let websocket = Socket.serve({ port: options.websocketPort });
+  let websocket = Socket.serve({ port: options.webSocketPort });
   let httpOrigin: Origin = {
     scheme: options.httpScheme,
     host: options.httpHost,
@@ -67,10 +55,12 @@ export let serve = async (
   };
 
   let state: {
-    latestBuildHash?: string;
-    buildHashChannel?: Channel.Type<void>;
     appServer?: execa.ExecaChildProcess;
+    manifest?: Manifest;
     prevManifest?: Manifest;
+    appReady?: Channel.Type<void>;
+    hdr?: Promise<Record<string, string>>;
+    prevLoaderHashes?: Record<string, string>;
   } = {};
 
   let bin = await detectBin();
@@ -80,9 +70,12 @@ export let serve = async (
       stdio: "pipe",
       env: {
         NODE_ENV: "development",
-        PATH: `${bin}:${process.env.PATH}`,
+        PATH:
+          bin + (process.platform === "win32" ? ";" : ":") + process.env.PATH,
         REMIX_DEV_HTTP_ORIGIN: stringifyOrigin(httpOrigin),
       },
+      // https://github.com/sindresorhus/execa/issues/433
+      windowsHide: false,
     });
 
     if (newAppServer.stdin)
@@ -100,8 +93,8 @@ export let serve = async (
               if (matches) {
                 for (let match of matches) {
                   let buildHash = match[1];
-                  if (buildHash === state.latestBuildHash) {
-                    state.buildHashChannel?.ok();
+                  if (buildHash === state.manifest?.version) {
+                    state.appReady?.ok();
                   }
                 }
               }
@@ -118,39 +111,37 @@ export let serve = async (
 
   let dispose = await Compiler.watch(
     {
-      config: patchPublicPath(initialConfig, httpOrigin),
+      config: initialConfig,
       options: {
         mode: "development",
         sourcemap: true,
         onWarning: warnOnce,
         devHttpOrigin: httpOrigin,
-        devWebsocketPort: options.websocketPort,
+        devWebSocketPort: options.webSocketPort,
       },
     },
     {
-      reloadConfig: async (root) => {
-        let config = await readConfig(root);
-        return patchPublicPath(config, httpOrigin);
-      },
-      onBuildStart: (ctx) => {
-        state.buildHashChannel?.err();
+      onBuildStart: async (ctx) => {
+        state.appReady?.err();
         clean(ctx.config);
         websocket.log(state.prevManifest ? "Rebuilding..." : "Building...");
+
+        state.hdr = HDR.detectLoaderChanges(ctx);
       },
-      onBuildFinish: async (ctx, durationMs, manifest) => {
-        if (!manifest) return;
+      onBuildManifest: (manifest: Manifest) => {
+        state.manifest = manifest;
+      },
+      onBuildFinish: async (ctx, durationMs, succeeded) => {
+        if (!succeeded) return;
 
         websocket.log(
           (state.prevManifest ? "Rebuilt" : "Built") +
             ` in ${prettyMs(durationMs)}`
         );
-        let prevManifest = state.prevManifest;
-        state.prevManifest = manifest;
-        state.latestBuildHash = manifest.version;
-        state.buildHashChannel = Channel.create();
+        state.appReady = Channel.create();
 
         let start = Date.now();
-        console.log(`Waiting for app server (${state.latestBuildHash})`);
+        console.log(`Waiting for app server (${state.manifest?.version})`);
         if (
           options.command &&
           (state.appServer === undefined || options.restart)
@@ -158,21 +149,30 @@ export let serve = async (
           await kill(state.appServer);
           state.appServer = startAppServer(options.command);
         }
-        let { ok } = await state.buildHashChannel.result;
+        let { ok } = await state.appReady.result;
         // result not ok -> new build started before this one finished. do not process outdated manifest
-        if (!ok) return;
-        console.log(`App server took ${prettyMs(Date.now() - start)}`);
+        let loaderHashes = await state.hdr;
+        if (ok) {
+          console.log(`App server took ${prettyMs(Date.now() - start)}`);
+          if (state.manifest && loaderHashes && state.prevManifest) {
+            let updates = HMR.updates(
+              ctx.config,
+              state.manifest,
+              state.prevManifest,
+              loaderHashes,
+              state.prevLoaderHashes
+            );
+            websocket.hmr(state.manifest, updates);
 
-        if (manifest.hmr && prevManifest) {
-          let updates = HMR.updates(ctx.config, manifest, prevManifest);
-          websocket.hmr(manifest, updates);
-
-          let hdr = updates.some((u) => u.revalidate);
-          console.log("> HMR" + (hdr ? " + HDR" : ""));
-        } else if (prevManifest !== undefined) {
-          websocket.reload();
-          console.log("> Live reload");
+            let hdr = updates.some((u) => u.revalidate);
+            console.log("> HMR" + (hdr ? " + HDR" : ""));
+          } else if (state.prevManifest !== undefined) {
+            websocket.reload();
+            console.log("> Live reload");
+          }
         }
+        state.prevManifest = state.manifest;
+        state.prevLoaderHashes = loaderHashes;
       },
       onFileCreated: (file) =>
         websocket.log(`File created: ${relativePath(file)}`),
@@ -184,19 +184,6 @@ export let serve = async (
   );
 
   let httpServer = express()
-    // statically serve built assets
-    .use((_, res, next) => {
-      res.header("Access-Control-Allow-Origin", "*");
-      next();
-    })
-    .use(
-      "/build",
-      express.static(initialConfig.assetsBuildDirectory, {
-        immutable: true,
-        maxAge: "1y",
-      })
-    )
-
     // handle `broadcastDevReady` messages
     .use(express.json())
     .post("/ping", (req, res) => {
@@ -205,8 +192,8 @@ export let serve = async (
         console.warn(`Unrecognized payload: ${req.body}`);
         res.sendStatus(400);
       }
-      if (buildHash === state.latestBuildHash) {
-        state.buildHashChannel?.ok();
+      if (buildHash === state.manifest?.version) {
+        state.appReady?.ok();
       }
       res.sendStatus(200);
     })
@@ -232,13 +219,15 @@ let relativePath = (file: string) => path.relative(process.cwd(), file);
 
 let kill = async (p?: execa.ExecaChildProcess) => {
   if (p === undefined) return;
-  // `execa`'s `kill` is not reliable on windows
+  let channel = Channel.create<void>();
+  p.on("exit", channel.ok);
+
+  // https://github.com/nodejs/node/issues/12378
   if (process.platform === "win32") {
     await execa("taskkill", ["/pid", String(p.pid), "/f", "/t"]);
-    return;
+  } else {
+    p.kill("SIGTERM", { forceKillAfterTimeout: 1_000 });
   }
 
-  // wait one tick of the event loop so that we guarantee app server gets killed before proceeding
-  p.kill("SIGTERM", { forceKillAfterTimeout: 0 });
-  await new Promise((resolve) => setTimeout(resolve, 0));
+  await channel.result;
 };